tripleo.sh: skip plan delete for mitaka
At the end of the overcloud delete, we delete the plan too.
The overcloud plan delete command was introduced only in the newton
cycle, so mitaka tests fail with an error.
This patch skip plan delete if we are testing overcloud delete in
mitaka.

Closes-Bug: 1657762

+++ scripts/tripleo.sh
@@ -769,12 +769,14 @@ function overcloud_delete {
        openstack stack show $OVERCLOUD_ID
        exit 1
     fi
-    openstack overcloud plan delete "$OVERCLOUD_NAME" && exitval=0 || exitval=1
-    if [ ${exitval} -eq 0 ]; then
-        log "Overcloud $OVERCLOUD_ID plan delete SUCCESS"
-    else
-        log "Overcloud $OVERCLOUD_ID plan delete FAILED"
-        exit 1
+    if [[ "${STABLE_RELEASE}" != "mitaka" ]] ; then
+        openstack overcloud plan delete "$OVERCLOUD_NAME" && exitval=0 || exitval=1
+        if [ ${exitval} -eq 0 ]; then
+            log "Overcloud $OVERCLOUD_ID plan delete SUCCESS"
+        else
+            log "Overcloud $OVERCLOUD_ID plan delete FAILED"
+            exit 1
+        fi
     fi
 }
